requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.7.1=pypi_0
apex=0.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
async-timeout=4.0.3=pypi_0
asynctest=0.13.0=pypi_0
attrs=23.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2022.12.7=py37h06a4308_0
cffi=1.15.1=py37h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
cloudpickle=2.2.1=pypi_0
comm=0.1.4=pypi_0
cryptography=39.0.1=py37h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
dearpygui=1.10.1=pypi_0
debugpy=1.7.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.6=py37h06a4308_0
einops=0.4.1=pypi_0
entrypoints=0.4=pypi_0
exceptiongroup=1.1.3=pypi_0
fastjsonschema=2.18.1=pypi_0
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
fonttools=4.38.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.59.0=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
idna=3.4=py37h06a4308_0
imageio=2.31.2=pypi_0
imageio-ffmpeg=0.4.9=pypi_0
importlib-metadata=6.7.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.16.2=pypi_0
ipython=7.34.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.1=py37h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=4.12.0=pypi_0
jupyter-server=1.24.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.6.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lpips=0.1.4=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.5.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.4=pypi_0
nbclassic=1.0.0=pypi_0
nbclient=0.7.4=pypi_0
nbconvert=7.6.0=pypi_0
nbformat=5.8.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
nettle=3.7.3=hbbd107a_1
notebook=6.5.6=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.2=pypi_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=py37h6a678d5_0
pip=22.3.1=py37h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=3.20.3=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydeprecate=0.3.2=pypi_0
pygments=2.16.1=pypi_0
pymcubes=0.1.4=pypi_0
pyopenssl=23.0.0=py37h06a4308_0
pyparsing=3.0.9=py37h06a4308_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=py37_1
python=3.7.16=h7a1cb2a_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-lightning=1.7.7=pypi_0
pytorch-mutex=1.0=cuda
pyyaml=6.0.1=pypi_0
pyzmq=24.0.1=pypi_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=py37h51133e4_1
scipy=1.7.3=py37h6c91a56_2
send2trash=1.8.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h5eee18b_0
tensorboard=2.11.2=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
terminado=0.17.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pypi_0
tinycudann=1.7=pypi_0
tk=8.6.12=h1ccaba5_0
torch-scatter=2.0.9=pypi_0
torchaudio=0.11.0=py37_cu113
torchmetrics=0.11.4=pypi_0
torchvision=0.12.0=py37_cu113
tornado=6.2=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pypi_0
trimesh=4.0.0=pypi_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.14=py37h06a4308_0
vren=2.0=pypi_0
wcwidth=0.2.8=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.1=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py37h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.15.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
